Propagate email and uuid through matching data

+++ fdbt-site/src/pages/thankyou.tsx
@@ -1,17 +1,21 @@
 import React, { ReactElement } from 'react';
 import { NextPageContext } from 'next';
+import { parseCookies } from 'nookies';
+import { decode } from 'jsonwebtoken';
 import TwoThirdsLayout from '../layout/Layout';
-import { FEEDBACK_LINK } from '../constants';
+import { FEEDBACK_LINK, ID_TOKEN_COOKIE } from '../constants';
 import { getUuidFromCookies, deleteAllCookiesOnServerSide } from '../utils';
+import { CognitoIdToken } from '../interfaces';
 
 const title = 'Thank You - Fares Data Build Tool';
 const description = 'Thank you page for the Fares Data Build Tool';
 
 type ThankYouProps = {
     uuid: string;
+    emailAddress: string;
 };
 
-const ThankYou = ({ uuid }: ThankYouProps): ReactElement => (
+const ThankYou = ({ uuid, emailAddress }: ThankYouProps): ReactElement => (
     <TwoThirdsLayout title={title} description={description}>
         <div className="govuk-panel govuk-panel--confirmation">
             <h1 className="govuk-panel__title" id="thank-you-page-heading">
@@ -26,9 +30,10 @@ const ThankYou = ({ uuid }: ThankYouProps): ReactElement => (
         <h2 className="govuk-heading-m">What happens next</h2>
         <p className="govuk-body">Thank you for submitting your fares data.</p>
         <p className="govuk-body">
-            Your data will be converted to the NeTEx format and published to the open data hub. You will be contacted
-            should there be a problem with your upload.
+            Your data will be converted to the NeTEx format and will be emailed to <strong>{emailAddress}</strong>{' '}
+            within the next 5 minutes for the fares you have just told us about.
         </p>
+        <p className="govuk-body">You will be contacted should there be a problem with your upload.</p>
         <p className="govuk-body">
             If you would like to provide feedback on this service, please click{' '}
             <a href={FEEDBACK_LINK} className="govuk-link">
@@ -46,8 +51,15 @@ export const getServerSideProps = (ctx: NextPageContext): {} => {
     const uuid = getUuidFromCookies(ctx);
     console.info('transaction complete', { uuid });
 
+    const cookies = parseCookies(ctx);
+    const idToken = cookies[ID_TOKEN_COOKIE];
+    const decodedIdToken = decode(idToken) as CognitoIdToken;
+    if (!decodedIdToken.email) {
+        throw new Error('Could not extract the user email address from their ID token');
+    }
+
     deleteAllCookiesOnServerSide(ctx);
-    return { props: { uuid } };
+    return { props: { uuid, emailAddress: decodedIdToken.email } };
 };
 
 export default ThankYou;
